+++ NotificationExtension/NotificationService.swift
@@ -46,13 +46,17 @@ class NotificationService: UNNotificationServiceExtension {
 			if let bestAttemptContent = bestAttemptContent {
 				let decodedMessage = try conversationTopic.toXMTP(client: client).decode(envelope)
 
+				var conversation = DB.Conversation.find(id: conversationTopic.conversationID)
+				conversation?.updatedAt = decodedMessage.sent
+				try conversation?.save()
+
 				// Don't notify when we're the sender
 				if decodedMessage.senderAddress == client.address {
 					contentHandler(UNNotificationContent())
 					return
 				}
 
-				if let conversation = DB.Conversation.find(id: conversationTopic.conversationID) {
+				if let conversation {
 					bestAttemptContent.title = conversation.title
 				}
 				bestAttemptContent.body = (try? decodedMessage.content()) ?? "no content"
