+++ xmtp-inbox-ios/Models/DB.swift
@@ -5,6 +5,7 @@
 //  Created by Pat Nakajima on 2/2/23.
 //
 
+import Combine
 import CryptoKit
 import Foundation
 import GRDB
@@ -19,6 +20,18 @@ class DB {
 	}
 
 	private static let shared = DB()
+
+	static var observer: DatabaseQueue {
+		shared.queue
+	}
+
+	static func observe<T>(setup: @escaping (Database) -> AnyPublisher<T, Error>) -> DatabasePublishers.Value<AnyPublisher<T, any Error>> {
+		let observation = ValueObservation.tracking { db in
+			setup(db)
+		}
+
+		return observation.publisher(in: shared.queue)
+	}
 
 	static func prepareTest(client: XMTP.Client) throws {
 		shared.mode = .test
