+++ xmtp-inbox-ios/Views/MessageListView.swift
@@ -23,30 +23,6 @@ class MessageTableViewCell: UITableViewCell {
 	}
 }
 
-class MessageObserver: TransactionObserver {
-	var callback: () -> Void
-
-	init(callback: @escaping () -> Void) {
-		self.callback = callback
-	}
-
-	func databaseDidCommit(_: GRDB.Database) {}
-	func databaseDidRollback(_: GRDB.Database) {}
-
-	func databaseDidChange(with _: GRDB.DatabaseEvent) {
-		callback()
-		stopObservingDatabaseChangesUntilNextTransaction()
-	}
-
-	func observes(eventsOfKind eventKind: DatabaseEventKind) -> Bool {
-		if case let .insert(tableName) = eventKind, tableName == "message" {
-			return true
-		} else {
-			return false
-		}
-	}
-}
-
 class MessagesTableViewController: UITableViewController {
 	var loader: MessageLoader
 	var cancellables = [AnyCancellable]()
